+++ LeafTipTriangulationLib/include/Camera.h
@@ -12,48 +12,48 @@
 class Camera
 {
 public:
-	Camera(const glm::vec3& eye, const glm::vec3& at, const glm::vec3& up);
+	Camera(const glm::dvec3& eye, const glm::dvec3& at, const glm::dvec3& up);
 
-	Camera(const glm::vec3& eye,
-		   const glm::vec3& at,
-		   const glm::vec3& up,
-		   float fovy,
-		   float aspectRatio,
-		   const glm::vec2& viewportSize);
+	Camera(const glm::dvec3& eye,
+		   const glm::dvec3& at,
+		   const glm::dvec3& up,
+		   double fovy,
+		   double aspectRatio,
+		   const glm::dvec2& viewportSize);
 
-	Camera(const glm::vec3& eye,
-		   const glm::vec3& at,
-		   const glm::vec3& up,
-		   const glm::mat4& matV,
-		   const glm::mat4& matP,
-		   const glm::vec2& viewportSize);
+	Camera(const glm::dvec3& eye,
+		   const glm::dvec3& at,
+		   const glm::dvec3& up,
+		   const glm::dmat4& matV,
+		   const glm::dmat4& matP,
+		   const glm::dvec2& viewportSize);
 
-	const glm::vec3& eye() const { return m_eye; }
-	const glm::vec3& at() const { return m_at; }
-	const glm::vec3& up() const { return m_up; }
+	const glm::dvec3& eye() const { return m_eye; }
+	const glm::dvec3& at() const { return m_at; }
+	const glm::dvec3& up() const { return m_up; }
 
-	const glm::mat4& matV() const { return m_matV; }
-	const glm::mat4& matP() const { return m_matP; }
-	const glm::mat4& mat() const { return m_mat; }
+	const glm::dmat4& matV() const { return m_matV; }
+	const glm::dmat4& matP() const { return m_matP; }
+	const glm::dmat4& mat() const { return m_mat; }
 
-	const glm::vec4& viewport() const { return m_viewport; }
+	const glm::dvec4& viewport() const { return m_viewport; }
 
-	glm::vec3 project(const glm::vec3& point) const;
+	glm::dvec3 project(const glm::dvec3& point) const;
 
-	glm::vec3 unProject(const glm::vec3& point) const;
+	glm::dvec3 unProject(const glm::dvec3& point) const;
 
-	glm::vec2 windowToViewport(const glm::vec2& point) const;
+	glm::dvec2 windowToViewport(const glm::dvec2& point) const;
 
 private:
-	glm::vec3 m_eye;
-	glm::vec3 m_at;
-	glm::vec3 m_up;
+	glm::dvec3 m_eye;
+	glm::dvec3 m_at;
+	glm::dvec3 m_up;
 
-	glm::mat4 m_matV;
-	glm::mat4 m_matP;
-	glm::mat4 m_mat;
+	glm::dmat4 m_matV;
+	glm::dmat4 m_matP;
+	glm::dmat4 m_mat;
 
-	glm::vec4 m_viewport;
+	glm::dvec4 m_viewport;
 };
 
 /**
@@ -61,14 +61,14 @@ private:
  * \param camera A camera
  * \return The maximum resolution of the camera 
  */
-float computeMaximumCameraResolution(const Camera& camera);
+double computeMaximumCameraResolution(const Camera& camera);
 
 /**
  * \brief Return the maximum resolution of a list of cameras
  * \param cameras A list of cameras
  * \return The maximum resolution of the list of cameras
  */
-float computeMaximumCameraResolution(const std::vector<Camera>& cameras);
+double computeMaximumCameraResolution(const std::vector<Camera>& cameras);
 
 /**
  * \brief Load camera from a list of files
@@ -84,4 +84,4 @@ std::vector<Camera> loadCamerasFromFiles(const std::vector<std::string>& files);
  * \return A list of rays from cameras
  */
 std::vector<std::vector<Ray>> computeRays(const std::vector<Camera>& cameras,
-							              const std::vector<std::vector<glm::vec2>>& points);
+							              const std::vector<std::vector<glm::dvec2>>& points);
